♻️ RJYWVG task: split finish spec helpers
